Create Khan Academy_Project1_Store
The project for this course was to make a fake store. It must include a table for all the items in my fake store, and at least 5 columns for the kind of data I'd need to store. Then I have to use “SELECT” statements to order my items by price and show at least one statistic about the items. After creating and adding items to my store, I found the sum of everything from the price column, which was 43.